@@ -5,18 +5,19 @@
     },
     "dependencies": {
         "@floating-ui/react-dom-interactions": "^0.6.6",
-        "@fortawesome/fontawesome-svg-core": "^1.2.32",
-        "@fortawesome/free-solid-svg-icons": "^5.15.1",
-        "@fortawesome/react-fontawesome": "^0.1.11",
+        "@fortawesome/fontawesome-svg-core": "^6.7.2",
+        "@fortawesome/free-solid-svg-icons": "^6.7.2",
+        "@fortawesome/react-fontawesome": "^0.2.2",
         "@headlessui/react": "^1.6.4",
         "@heroicons/react": "^1.0.6",
         "@hot-loader/react-dom": "^16.14.0",
         "@preact/signals-react": "^1.2.1",
         "@tailwindcss/forms": "^0.5.2",
         "@tailwindcss/line-clamp": "^0.4.0",
+        "@types/chart.js": "^2.9.41",
         "axios": "^0.27.2",
         "boring-avatars": "^1.7.0",
-        "chart.js": "^3.8.0",
+        "chart.js": "3.8.0",
         "classnames": "^2.3.1",
         "codemirror": "^5.57.0",
         "copy-to-clipboard": "^3.3.1",
@@ -31,13 +32,13 @@
         "i18next-http-backend": "^1.4.1",
         "i18next-multiload-backend-adapter": "^1.0.0",
         "qrcode.react": "^1.0.1",
-        "react": "^16.14.0",
+        "react": "^19.0.0",
         "react-chartjs-2": "^4.2.0",
-        "react-dom": "npm:@hot-loader/react-dom",
+        "react-dom": "^19.0.0",
         "react-fast-compare": "^3.2.0",
         "react-hot-loader": "^4.12.21",
         "react-i18next": "^11.2.1",
-        "react-router-dom": "^5.1.2",
+        "react-router-dom": "^7.4.0",
         "react-transition-group": "^4.4.1",
         "reaptcha": "^1.7.2",
         "sockette": "^2.0.6",
@@ -78,12 +79,12 @@
         "@types/jest": "^28.1.3",
         "@types/node": "^14.11.10",
         "@types/qrcode.react": "^1.0.1",
-        "@types/react": "^16.14.0",
+        "@types/react": "17.0.2",
         "@types/react-copy-to-clipboard": "^4.3.0",
-        "@types/react-dom": "^16.9.16",
+        "@types/react-dom": "17.0.2",
         "@types/react-redux": "^7.1.1",
-        "@types/react-router": "^5.1.3",
-        "@types/react-router-dom": "^5.1.3",
+        "@types/react-router": "5.1.20",
+        "@types/react-router-dom": "5.3.3",
         "@types/react-transition-group": "^4.4.0",
         "@types/styled-components": "^5.1.7",
         "@types/uuid": "^3.4.5",
