@@ -1,29 +1,26 @@
 import React from 'react';
-import { Route, Switch, useRouteMatch } from 'react-router-dom';
+import { Route, Routes, useNavigate } from 'react-router-dom';
 import LoginContainer from '@/components/auth/LoginContainer';
+import LoginCheckpointContainer from '@/components/auth/LoginCheckpointContainer';
 import ForgotPasswordContainer from '@/components/auth/ForgotPasswordContainer';
 import ResetPasswordContainer from '@/components/auth/ResetPasswordContainer';
-import LoginCheckpointContainer from '@/components/auth/LoginCheckpointContainer';
 import { NotFound } from '@/components/elements/ScreenBlock';
-import { useHistory, useLocation } from 'react-router';
 
-export default () => {
-    const history = useHistory();
-    const location = useLocation();
-    const { path } = useRouteMatch();
+const AuthenticationRouter: React.FC = () => {
+    const navigate = useNavigate();
 
     return (
-        <div className={'pt-8 xl:pt-32'}>
-            <Switch location={location}>
-                <Route path={`${path}/login`} component={LoginContainer} exact />
-                <Route path={`${path}/login/checkpoint`} component={LoginCheckpointContainer} />
-                <Route path={`${path}/password`} component={ForgotPasswordContainer} exact />
-                <Route path={`${path}/password/reset/:token`} component={ResetPasswordContainer} />
-                <Route path={`${path}/checkpoint`} />
-                <Route path={'*'}>
-                    <NotFound onBack={() => history.push('/auth/login')} />
-                </Route>
-            </Switch>
+        <div className="pt-8 xl:pt-32">
+            <Routes>
+                <Route path="/auth/login" element={<LoginContainer />} />
+                <Route path="/auth/login/checkpoint" element={<LoginCheckpointContainer />} />
+                <Route path="/auth/password" element={<ForgotPasswordContainer />} />
+                <Route path="/auth/password/reset/:token" element={<ResetPasswordContainer />} />
+                <Route path="/auth/checkpoint" element={<></>} />
+                <Route path="*" element={<NotFound onBack={() => navigate('/auth/login')} />} />
+            </Routes>
         </div>
     );
 };
+
+export default AuthenticationRouter;
