@@ -1,29 +1,29 @@
 import React from 'react';
-import { Route } from 'react-router';
+import { Route, useLocation } from 'react-router-dom';
 import { SwitchTransition } from 'react-transition-group';
+import styled from 'styled-components';
 import Fade from '@/components/elements/Fade';
-import styled from 'styled-components/macro';
 import tw from 'twin.macro';
 
-const StyledSwitchTransition = styled(SwitchTransition)`
+const StyledTransitionWrapper = styled.div`
     ${tw`relative`};
 
     & section {
-        ${tw`absolute w-full top-0 left-0`};
+        ${tw`relative w-full`};
     }
 `;
 
 const TransitionRouter: React.FC = ({ children }) => {
+    const location = useLocation();
+
     return (
-        <Route
-            render={({ location }) => (
-                <StyledSwitchTransition>
-                    <Fade timeout={150} key={location.pathname + location.search} in appear unmountOnExit>
-                        <section>{children}</section>
-                    </Fade>
-                </StyledSwitchTransition>
-            )}
-        />
+        <StyledTransitionWrapper>
+            <SwitchTransition>
+                <Fade timeout={150} key={location.pathname + location.search} in appear unmountOnExit>
+                    <section>{children}</section>
+                </Fade>
+            </SwitchTransition>
+        </StyledTransitionWrapper>
     );
 };
 
